+++ test/tracer_test.dart
@@ -1,5 +1,3 @@
-import 'dart:async';
-
 import 'package:tracer/tracer.dart';
 import 'package:test/test.dart';
 
@@ -54,40 +52,35 @@ class TracerTestFatal extends TracerHandler {
 }
 
 void main() {
-  test("Debug", () async {
+  test("Debug", () {
     expect(ranDebug, isFalse);
     final t = Tracer("testing_debug",
         logLevel: TracerLevel.debug, handlers: [TracerTestDebug()]);
     t.debug("This is a test debug message");
-    await Future.delayed(Duration(milliseconds: 100));
     expect(ranDebug, isTrue);
   });
-  test("Info", () async {
+  test("Info", () {
     expect(ranInfo, isFalse);
     final t = Tracer("testing_info", handlers: [TracerTestInfo()]);
     t.info("This is a test info message");
-    await Future.delayed(Duration(milliseconds: 1));
     expect(ranInfo, isTrue);
   });
-  test("Warn", () async {
+  test("Warn", () {
     expect(ranWarn, isFalse);
     final t = Tracer("testing_warn", handlers: [TracerTestWarn()]);
     t.warn("This is a test warn message");
-    await Future.delayed(Duration(milliseconds: 1));
     expect(ranWarn, isTrue);
   });
-  test("Error", () async {
+  test("Error", () {
     expect(ranError, isFalse);
     final t = Tracer("testing_error", handlers: [TracerTestError()]);
     t.error("This is a test error message");
-    await Future.delayed(Duration(milliseconds: 1));
     expect(ranError, isTrue);
   });
-  test("Fatal", () async {
+  test("Fatal", () {
     expect(ranFatal, isFalse);
     final t = Tracer("testing_fatal", handlers: [TracerTestFatal()]);
     t.fatal("This is a test fatal message");
-    await Future.delayed(Duration(milliseconds: 1));
     expect(ranFatal, isTrue);
   });
 }